chore(sdk-core): refactor Environment to core
Ticket: BG-45153
