+++ modules/bitgo/test/v2/unit/trading/tradingAccount.ts
@@ -4,7 +4,7 @@ import * as nock from 'nock';
 import { Enterprise } from '../../../../src/v2/enterprise';
 import { Wallet } from '../../../../src/v2/wallet';
 import { TestBitGo } from '../../../lib/test_bitgo';
-import { Environments } from '../../../../src/common';
+import { Environments } from '@bitgo/sdk-core';
 
 describe('Trading Accounts', function () {
   const microservicesUri = Environments['mock'].uri;
