+++ modules/sdk-core/src/index.ts
@@ -1,1 +1,1 @@
-export const hello = 1;
+export * from './types';
