+++ modules/sdk-core/src/types/environmentName.ts
@@ -1,0 +1,15 @@
+export type EnvironmentName =
+  | 'prod'
+  | 'staging'
+  | 'test'
+  | 'dev'
+  | 'latest'
+  | 'local'
+  | 'localNonSecure'
+  | 'mock'
+  | 'adminProd'
+  | 'adminTest'
+  | 'adminDev'
+  | 'adminLatest'
+  | 'custom'
+  | 'branch';
